requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
elasticsearch>=8.5.1
django-elasticsearch-dsl
django-elasticsearch-dsl-drf
elasticsearch-dsl
python-dotenv
